* commit '43ae13453126e98416638085b4dbffa685536c44':
  模板锁定问题
  bug

+++ fbi_web_report/src/com/fr/bi/web/report/services/BITemplateFolderDeleteAction.java
@@ -1,13 +1,16 @@
 package com.fr.bi.web.report.services;
 
+import com.fr.bi.cal.analyze.session.BISessionUtils;
 import com.fr.bi.fs.*;
 import com.fr.fs.control.UserControl;
 import com.fr.fs.web.service.ServiceUtils;
+import com.fr.json.JSONArray;
 import com.fr.web.core.ActionNoSessionCMD;
 import com.fr.web.utils.WebUtils;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -31,32 +34,54 @@ public class BITemplateFolderDeleteAction extends ActionNoSessionCMD {
         String sType = WebUtils.getHTTPRequestParameter(req, "type");
         long userId = ServiceUtils.getCurrentUserID(req);
         int type = Integer.parseInt(sType);
+        List<String> lockedList = new ArrayList<String>();
         switch (type){
             case DELETE_FOLDER:
-                removeFolderById(userId, id);
+                lockedList = removeFolderById(userId, id);
                 break;
             case DELETE_REPORT:
-                removeReportById(userId, Long.parseLong(id));
+                String userName = removeReportById(userId, Long.parseLong(id));
+                if(userName != null) {
+                    lockedList.add(userName);
+                }
                 break;
         }
+        JSONArray userList = new JSONArray();
+        for(int i = 0; i < lockedList.size(); i++) {
+            userList.put(lockedList.get(i));
+        }
+        WebUtils.printAsJSON(res, userList);
     }
 
-    private void removeReportById(long userId, long reportId) throws Exception {
-        BIDAOUtils.deleteBIReportById(userId, reportId);
-        UserControl.getInstance().getOpenDAO(BISharedReportDAO.class).removeSharedByReport(reportId, userId);
+    private String removeReportById(long userId, long reportId) throws Exception {
+        String userName = BISessionUtils.getCurrentEditingUserByReport(reportId, userId);
+        if(userName == null) {
+            BIDAOUtils.deleteBIReportById(userId, reportId);
+            UserControl.getInstance().getOpenDAO(BISharedReportDAO.class).removeSharedByReport(reportId, userId);
+        }
+        return userName;
     }
 
-    private void removeFolderById(long userId, String folderId) throws Exception {
-        HSQLBITemplateFolderDAO.getInstance().deleteByFolderID(folderId);
+    private List<String> removeFolderById(long userId, String folderId) throws Exception {
         List<BIReportNode> reports = BIDAOUtils.findByParentID(userId, folderId);
         List<BITemplateFolderNode> folders = HSQLBITemplateFolderDAO.getInstance().findTemplateFolderByParentId(folderId);
-        for(int i = 0; i < reports.size(); i++){
-            removeReportById(userId, reports.get(i).getId());
+        List<String> lockedList = new ArrayList<String>();
+        if(reports != null) {
+            for(int i = 0; i < reports.size(); i++){
+                String userName = removeReportById(userId, reports.get(i).getId());
+                if(userName != null) {
+                    lockedList.add(userName);
+                }
+            }
         }
         for(int i = 0; i < folders.size(); i++){
             String fId = folders.get(i).getFolderId();
-            HSQLBITemplateFolderDAO.getInstance().deleteByFolderID(fId);
-            removeFolderById(userId, fId);
+            List<String> singleList = removeFolderById(userId, fId);
+            lockedList.addAll(singleList);
         }
+        if(lockedList.size() == 0) {
+            HSQLBITemplateFolderDAO.getInstance().deleteByFolderID(folderId);
+        }
+        return lockedList;
     }
 }
